[PIO-162] ASF compliance
