+++ client/pom.xml
@@ -2,16 +2,15 @@
   <modelVersion>4.0.0</modelVersion>
 
   <parent>
-    <groupId>io.prediction</groupId>
-    <artifactId>sdk</artifactId>
-    <relativePath>../pom.xml</relativePath>
-    <version>0.9.6-SNAPSHOT</version>
+    <groupId>org.apache.predictionio</groupId>
+    <artifactId>sdk-java</artifactId>
+    <version>0.13.0-SNAPSHOT</version>
   </parent>
 
-  <artifactId>client</artifactId>
+  <artifactId>sdk-java-client</artifactId>
   <packaging>jar</packaging>
 
-  <name>PredictionIO Java SDK: Client</name>
+  <name>Apache PredictionIO Java SDK: Client</name>
 
   <properties>
     <project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
@@ -58,7 +57,6 @@
       <plugin>
         <groupId>org.apache.maven.plugins</groupId>
         <artifactId>maven-compiler-plugin</artifactId>
-        <version>3.0</version>
         <configuration>
           <encoding>UTF-8</encoding>
           <source>1.7</source>
@@ -68,7 +66,6 @@
       <plugin>
         <groupId>org.apache.maven.plugins</groupId>
         <artifactId>maven-source-plugin</artifactId>
-        <version>2.2.1</version>
         <executions>
           <execution>
             <id>attach-sources</id>
@@ -81,7 +78,6 @@
       <plugin>
         <groupId>org.apache.maven.plugins</groupId>
         <artifactId>maven-javadoc-plugin</artifactId>
-        <version>2.9</version>
         <executions>
           <execution>
             <id>attach-javadocs</id>
