--- client/src/main/java/io/prediction/package-info.java
@@ -1,10 +1,0 @@
-/**
- * This package contains classes that provide convenient access of the PredictionIO RESTful API.
- *
- * To create an app and perform predictions, please download the PredictionIO suite from http://prediction.io.
- *
- * Most functionality is provided by the DataClient and EngineClient class.
- *
- * @author The PredictionIO Team (<a href="http://prediction.io">http://prediction.io</a>)
- */
-package io.prediction;
