+++ client/src/main/java/org/apache/predictionio/sdk/java/DateTimeAdapter.java
@@ -1,4 +1,21 @@
-package io.prediction;
+/*
+ * Licensed to the Apache Software Foundation (ASF) under one or more
+ * contributor license agreements.  See the NOTICE file distributed with
+ * this work for additional information regarding copyright ownership.
+ * The ASF licenses this file to You under the Apache License, Version 2.0
+ * (the "License"); you may not use this file except in compliance with
+ * the License.  You may obtain a copy of the License at
+ *
+ *    http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package org.apache.predictionio.sdk.java;
 
 import com.google.gson.JsonDeserializationContext;
 import com.google.gson.JsonDeserializer;
@@ -15,7 +32,6 @@ import java.lang.reflect.Type;
 /**
  * DateTimeAdapter turns a String in ISO 8601 format into a DateTime object, and vice versa.
  *
- * @author The PredictionIO Team (<a href="http://prediction.io">http://prediction.io</a>)
  * @version 0.8.3
  * @since 0.8.0
  */
