+++ client/src/main/java/org/apache/predictionio/sdk/java/EngineClient.java
@@ -1,9 +1,25 @@
-package io.prediction;
+/*
+ * Licensed to the Apache Software Foundation (ASF) under one or more
+ * contributor license agreements.  See the NOTICE file distributed with
+ * this work for additional information regarding copyright ownership.
+ * The ASF licenses this file to You under the Apache License, Version 2.0
+ * (the "License"); you may not use this file except in compliance with
+ * the License.  You may obtain a copy of the License at
+ *
+ *    http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package org.apache.predictionio.sdk.java;
 
 import com.google.gson.Gson;
 import com.google.gson.GsonBuilder;
 import com.google.gson.JsonObject;
-import com.ning.http.client.Request;
 import com.ning.http.client.RequestBuilder;
 
 import org.joda.time.DateTime;
@@ -16,7 +32,6 @@ import java.util.concurrent.ExecutionException;
  * EngineClient contains generic methods sendQuery() and sendQueryAsFuture()
  * for sending queries.
  *
- * @author The PredictionIO Team (<a href="http://prediction.io">http://prediction.io</a>)
  * @version 0.8.3
  * @since 0.8.0
  */
@@ -113,7 +128,7 @@ public class EngineClient extends BaseClient {
         int status = response.get().getStatus();
         String message = response.get().getMessage();
 
-        if (status != BaseClient.HTTP_OK) {
+        if (status != HTTP_OK) {
             throw new IOException(status + " " + message);
         }
         return ((JsonObject) parser.parse(message));
