+++ pom.xml
@@ -1,18 +1,18 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
   <modelVersion>4.0.0</modelVersion>
-  <groupId>io.prediction</groupId>
-  <artifactId>sdk</artifactId>
-  <version>0.9.6-SNAPSHOT</version>
-  <url>http://prediction.io</url>
+  <groupId>org.apache.predictionio</groupId>
+  <artifactId>sdk-java</artifactId>
+  <version>0.13.0-SNAPSHOT</version>
+  <url>http://predictionio.apache.org</url>
   <packaging>pom</packaging>
-  <name>PredictionIO Java SDK</name>
-  <description>The PredictionIO Java SDK includes an API client and sample code.</description>
+  <name>Apache PredictionIO Java SDK</name>
+  <description>The Apache PredictionIO Java SDK includes an API client and sample code.</description>
 
   <parent>
-    <groupId>org.sonatype.oss</groupId>
-    <artifactId>oss-parent</artifactId>
-    <version>7</version>
+    <groupId>org.apache</groupId>
+    <artifactId>apache</artifactId>
+    <version>21</version>
   </parent>
 
   <dependencies>
@@ -23,34 +23,12 @@
     </dependency>
   </dependencies>
 
-  <licenses>
-    <license>
-      <name>The Apache Software License, Version 2.0</name>
-      <url>http://www.apache.org/licenses/LICENSE-2.0.txt</url>
-      <distribution>repo</distribution>
-    </license>
-  </licenses>
-
-  <developers>
-    <developer>
-      <id>predictionio</id>
-      <name>The PredictionIO Team</name>
-      <email>support@prediction.io</email>
-    </developer>
-  </developers>
-
   <contributors>
     <contributor>
       <name>Cong Qin</name>
       <email>Cong.Charlie.Qin@gmail.com</email>
     </contributor>
   </contributors>
-
-  <scm>
-    <connection>scm:git:git@github.com:PredictionIO/PredictionIO-Java-SDK.git</connection>
-    <developerConnection>scm:git:git@github.com:PredictionIO/PredictionIO-Java-SDK.git</developerConnection>
-    <url>git@github.com:PredictionIO/PredictionIO-Java-SDK.git</url>
-  </scm>
 
   <properties>
     <project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
@@ -61,54 +39,4 @@
   <modules>
     <module>client</module>
   </modules>
-
-  <distributionManagement>
-    <snapshotRepository>
-      <id>ossrh</id>
-      <url>https://oss.sonatype.org/content/repositories/snapshots</url>
-    </snapshotRepository>
-    <repository>
-      <id>ossrh</id>
-      <url>https://oss.sonatype.org/service/local/staging/deploy/maven2/</url>
-    </repository>
-  </distributionManagement>
-
-  <build>
-    <plugins>
-      <plugin>
-        <groupId>org.apache.maven.plugins</groupId>
-        <artifactId>maven-compiler-plugin</artifactId>
-        <version>3.0</version>
-        <configuration>
-          <encoding>UTF-8</encoding>
-          <source>1.7</source>
-          <target>1.7</target>
-        </configuration>
-      </plugin>
-      <plugin>
-        <groupId>org.apache.maven.plugins</groupId>
-        <artifactId>maven-gpg-plugin</artifactId>
-        <executions>
-          <execution>
-            <id>sign-artifacts</id>
-            <phase>verify</phase>
-            <goals>
-              <goal>sign</goal>
-            </goals>
-          </execution>
-        </executions>
-      </plugin>
-      <plugin>
-        <groupId>org.sonatype.plugins</groupId>
-        <artifactId>nexus-staging-maven-plugin</artifactId>
-        <version>1.6</version>
-        <extensions>true</extensions>
-        <configuration>
-          <serverId>ossrh</serverId>
-          <nexusUrl>https://oss.sonatype.org/</nexusUrl>
-        </configuration>
-      </plugin>
-    </plugins>
-  </build>
-
 </project>
